requirements.txt
streamlit==1.31.1
python-dotenv==1.0.0
PyPDF2==3.0.1
langchain==0.1.11
langchain_community==0.0.27
langchain_openai==0.0.8
transformers==4.38.2
tiktoken==0.1.2
python-docx==1.1.0